add guard to prevent FLUSHDB and FLUSHALL execution on readonly replicas

+++ libs/server/Resp/BasicCommands.cs
@@ -892,6 +892,13 @@ namespace Garnet.server
                 return AbortWithWrongNumberOfArguments(nameof(RespCommand.FLUSHDB));
             }
 
+            if (storeWrapper.serverOptions.EnableCluster && storeWrapper.clusterProvider.IsReplica() && !clusterSession.ReadWriteSession)
+            {
+                while (!RespWriteUtils.TryWriteError(CmdStrings.RESP_ERR_FLUSHALL_READONLY_REPLICA, ref dcurr, dend))
+                    SendAndReset();
+                return true;
+            }
+
             FlushDb(RespCommand.FLUSHDB);
 
             return true;
@@ -905,6 +912,13 @@ namespace Garnet.server
             if (parseState.Count > 2)
             {
                 return AbortWithWrongNumberOfArguments(nameof(RespCommand.FLUSHALL));
+            }
+
+            if (storeWrapper.serverOptions.EnableCluster && storeWrapper.clusterProvider.IsReplica() && !clusterSession.ReadWriteSession)
+            {
+                while (!RespWriteUtils.TryWriteError(CmdStrings.RESP_ERR_FLUSHALL_READONLY_REPLICA, ref dcurr, dend))
+                    SendAndReset();
+                return true;
             }
 
             // Since Garnet currently only supports a single database,
